This is the Great MUSTard Cleanup.  All uses of the RFC terms "must", "must not", "shall", "shall not", "should", "should not", "may", "required", "recommended", and "optional" now conform to RFC 2119. In particular, many uses of "must" have been replaced by "it is an error unless" or something similar, as they are not absolute restrictions on all implementations.  Some uses of "may" have become "can" or "might". Other uses of "may" and "should" have simply been dropped: for example, rather than speaking of bindings which may surround a macro, we simply speak of bindings which surround a macro.
All uses of "may not", except in the context "may or may not", have been
flushed, as they are readily confused with "must not".


The following additional changes were made on the fly:

The ``exact-complex`` feature was added to Appendix B.

Obsolete \nodomain comments, many of which contained bad MUSTard, have
been removed.

The phrase "The \var{port} argument" has become simply "\var{Port}"
to conform with existing uses.