* Replaced the word "module" with the word "library" throughout the notes section, which was the only section of the document that used the old term.
* Added \rthreers, \rfourrs, \rfivers, \rsixrs, and \rsevenrs and used
them to maintain the conventional superscripting in the Scheme
reports' names throughout the document.

* Added \sharpfalse and \sharptrue for #false and #true and used them.

* In notes, fixed capitalization of identifiers that appeared at the
beginnings of sentences.  They had been changed to capitals, but
that's inconsistent with what I've been doing in the rest of the
document and is technically incorrect given that the language is now
case sensitive, so I reworded entries to move the identifiers out of
first position.

* Added a bibliographic entry for RFC 2119.  Fixed the punctuation of
the SRFI-45 entry.

* Clarified wording, improved formatting, and fixed typos,
punctuation, and grammar errors.

* Checked in the latest generated "stdmod.tex".