The formal semantics define the interactions between call/cc and dynamic-wind, not parameterize or exceptions which are simply defined in terms of dynamic-wind.  Being hard to understand is subjective and not particularly relevant.  The memory issue is more complicated than consuming an amount of memory depending on active continuations.
Overall, let's not bury call/cc until we have a replacement.  The
appropriate way to handle this is in the large language, when we
introduce shift+reset, to discuss in more detail the problems with
call/cc as motivation for shift+reset.  In the small language it
suffices to point out that raise+guard can be used for more structured
non-local exits, since raise+guard are already defined.